fixes for transaction dialog
git-svn-id: https://svn.code.sf.net/p/moneymanagerex/code/trunk@4980 7f5d68d0-c712-0410-839f-9b116a576cc5

+++ src/splittransactionsdialog.cpp
@@ -189,7 +189,7 @@ void SplitTransactionDialog::OnOk( wxCommandEvent& /*event*/ )
 
 void SplitTransactionDialog::OnButtonRemoveClick( wxCommandEvent& event )
 {
-    if (selectedIndex_ < 0 || selectedIndex_ >= this->m_local_splits.size()) return;
+    if (selectedIndex_ < 0 || selectedIndex_ >= (int)this->m_local_splits.size()) return;
     this->m_local_splits.erase(this->m_local_splits.begin() + selectedIndex_);
     DataToControls();
 }
